backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifultable==0.8.0
boto3==1.26.77
botocore==1.29.77
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Django==4.1.4
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fandogh-cli==1.26.10
idna==3.4
jmespath==1.0.1
kavenegar==1.1.2
Pillow==9.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==5.4.1
requests==2.28.2
requests-toolbelt==0.10.1
retrying==1.3.4
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzlocal==2.1
urllib3==1.26.14
websocket-client==1.5.1
whitenoise==6.3.0
